rewire the views
