+++ src/App.vue
@@ -111,19 +111,9 @@ export default {
     showSnackbar: false,
     items: [
       {
-        name: 'Add stream',
-        icon: '➕',
-        to: '/add'
-      },
-      {
         name: 'Streams',
         icon: '📃',
         to: '/'
-      },
-      {
-        name: 'SingleStreamDev',
-        icon: '📃',
-        to: '/singleStream'
       },
       {
         name: 'Log out',
